# public/docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Keyp.fo API",
    "version": "1.0.0",
    "description": "API for Keyp.fo services including stores, products, wishlists, and gift ideas. For detailed implementation instructions, see the guide above.",
    "x-implementation-guide": "/docs/implementation-guide.md"
  },
  "servers": [
    {
      "url": "https://api.keyp.fo/v1",
      "description": "Production API"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Firebase JWT token obtained from Firebase Authentication"
      }
    },
    "schemas": {
...
      "Product": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "storeId": { "type": "string" },
          "name": { "type": "string" },
          "description": { "type": "string" },
          "price": { "type": "number" },
          "category": { "type": "string" },
          "inStock": { "type": "boolean" },
          "imageUrl": { "type": "string" }
        }
      },
      "Wishlist": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "userId": { "type": "string" },
          "name": { "type": "string" },
          "description": { "type": "string" },
          "isPublic": { "type": "boolean" },
          "items": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "productId": { "type": "string" },
                "storeId": { "type": "string" },
                "priority": { "type": "integer", "enum": [1, 2, 3] },
                "notes": { "type": "string" },
                "addedAt": { "type": "string", "format": "date-time" },
                "quantity": { "type": "integer", "minimum": 1 }
              }
            }
          },
          "viewers": {
            "type": "array",
            "items": { "type": "string" }
          },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": { "type": "string" },
              "message": { "type": "string" },
              "details": {
                "type": "object",
                "properties": {
                  "field": { "type": "string" },
                  "reason": { "type": "string" }
                }
              }
            }
          }
        }
      }
    }
  },
  "paths": {
...
    "/products": {
      "get": {
        "summary": "List products",
        "parameters": [
          {
            "name": "storeId",
            "in": "query",
            "schema": { "type": "string" }
          },
          {
            "name": "category",
            "in": "query",
            "schema": { "type": "string" }
          },
          {
            "name": "minPrice",
            "in": "query",
            "schema": { "type": "number" }
          },
          {
            "name": "maxPrice",
            "in": "query",
            "schema": { "type": "number" }
          },
          {
            "name": "inStock",
            "in": "query",
            "schema": { "type": "boolean" }
          }
        ],
        "responses": {
          "200": {
            "description": "List of products",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Product" }
                }
              }
            }
          }
        }
      }
    },
    "/wishlists": {
      "get": {
        "summary": "List user's wishlists",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "List of wishlists",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Wishlist" }
                },
                "example": {
                  "wishlists": [
                    {
                      "id": "abc123",
                      "name": "Birthday Wishlist",
                      "description": "My birthday wishes for 2024",
                      "isPublic": true,
                      "items": [
                        {
                          "productId": "prod123",
                          "storeId": "store456",
                          "priority": 1,
                          "notes": "Blue color preferred",
                          "quantity": 1
                        }
                      ]
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid or missing token",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new wishlist",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name"],
                "properties": {
                  "name": { "type": "string" },
                  "description": { "type": "string" },
                  "isPublic": { "type": "boolean", "default": false },
                  "items": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "required": ["productId", "storeId"],
                      "properties": {
                        "productId": { "type": "string" },
                        "storeId": { "type": "string" },
                        "priority": { "type": "integer", "enum": [1, 2, 3] },
                        "notes": { "type": "string" },
                        "quantity": { "type": "integer", "minimum": 1 }
                      }
                    }
                  }
                }
              },
              "example": {
                "name": "Birthday Wishlist",
                "description": "My birthday wishes for 2024",
                "isPublic": true,
                "items": [
                  {
                    "productId": "prod123",
                    "storeId": "store456",
                    "priority": 1,
                    "notes": "Blue color preferred",
                    "quantity": 1
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Wishlist created successfully",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Wishlist" }
              }
            }
          },
          "400": {
            "description": "Invalid request body",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid or missing token",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
} 